+++ dts-admin-dao/src/main/java/com/utopia/data/transfer/admin/dao/mapper/base/PairBeanRepository.java
@@ -1,9 +1,5 @@
 package com.utopia.data.transfer.admin.dao.mapper.base;
 
-import com.utopia.data.transfer.admin.dao.base.BaseDao;
-import com.utopia.data.transfer.admin.dao.entity.PairBean;
-import com.utopia.data.transfer.admin.dao.entity.PairBeanDal;
-import com.utopia.data.transfer.admin.dao.entity.PairDetail;
 import com.utopia.data.transfer.admin.dao.mapper.PairBeanMapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -11,5 +7,5 @@ import java.util.List;
 
 public interface PairBeanRepository extends PairBeanMapper {
 
-    List<PairDetail> selectByPipelineId(@Param("id") Integer pipelineId);
+    List<PairDetail> selectByPipelineId(@Param("id") Long pipelineId);
 }
