+++ dts-admin-dao/src/main/java/com/utopia/data/transfer/admin/dao/mapper/base/RegionBeanRepository.java
@@ -11,5 +11,5 @@ import java.util.List;
 public interface RegionBeanRepository extends RegionBeanMapper {
 
 
-    List<RegionBean> selectByPipelineId(@Param("id") Integer pipelineId);
+    List<RegionBean> selectByPipelineId(@Param("id") Long pipelineId);
 }
